+++ src/tools/views/getViewImage.ts
